+++ backend/app/api/v1/endpoints/users.py
@@ -1,9 +1,12 @@
 from typing import Any, List
 
-from fastapi import APIRouter, HTTPException, Request
+from fastapi import APIRouter, HTTPException, Request, status, Body
 from api.deps import auth_guard
-from crud import crud_users
+
 from mongo.schemas.users import *
+from crud.crud_users import users
+from mongo.models.users import User
+from mongo.schemas.users import UserUpdateProfile
 
 router = APIRouter()
 
@@ -11,74 +14,93 @@ router = APIRouter()
 @router.get("/", response_model=User)
 @auth_guard("admin")
 def get_users(request: Request):
-    """
-    Retrieve users.
-    """
-    try:
-        crud_users.CRUD_users.get_all()
-        return {"users": "user", "use": "use"}
-    except HTTPException:
-        pass
+	"""_summary_
+
+	Args:
+		request (Request): _description_
+
+	Raises:
+		HTTPException: _description_
+
+	Returns:
+		_type_: _description_
+	"""
+	users_list = users.get_all()
+	if len(users_list) is 0:
+		raise HTTPException(
+			status_code=status.HTTP_404_NOT_FOUND,
+			detail="No users were found",
+		)
+	return users_list
 
 
 @router.get("/me", response_model=User)
 @auth_guard("user")
 def get_my_user(request: Request):
-    """
-    Retrieve own user.
-    """
-    try:
-        crud_users.CRUD_users.get_my()
-    except HTTPException:
-        pass
+	"""_summary_
+
+	Args:
+		request (Request): _description_
+
+	Returns:
+		_type_: _description_
+	"""
+	return request.attach_user
 
 
-@router.get("/{id}", response_model=User)
+@router.get("/{id}")
+def get_user_by_id(request: Request, id: str):
+	"""_summary_
+
+	Args:
+		request (Request): _description_
+		id (str): _description_
+
+	Raises:
+		HTTPException: _description_
+
+	Returns:
+		_type_: _description_
+	"""
+	user = users.get(id)
+	if not user:
+		raise HTTPException(
+			status_code=status.HTTP_404_NOT_FOUND,
+			detail="User not found"
+		)
+	
+	user.id = str(user.id)
+	return user.__dict__
+
+
+@router.patch("/me", response_model=UserUpdate)
 @auth_guard("user")
-def get_user_by_id(request: Request):
-    """
-    Retrieve a specified user.
-    """
-    try:
-        # crud_users.CRUD_users.get_by_id(id)
-        print(request.attach_user)
-    except HTTPException:
-        pass
+def update_user(request: Request, update_data: UserUpdateProfile = Body(...)):
+	"""_summary_
+
+	Args:
+		request (Request): _description_
+		update_data (UserUpdateProfile, optional): _description_. Defaults to Body(...).
+
+	Returns:
+		_type_: _description_
+	"""
+	user = User.assert_model_id(request.attach_user)
+	users.update(user, update_data)
+	return { "status": "OK" }
 
 
-@router.post("/", response_model=UserCreate)
-def create_user(request: Request):
-    """
-    Create a new user.
-    """
-    try:
-        crud_users.CRUD_users.create()
-    except HTTPException:
-        pass
-
-
-@router.patch("/{id}", response_model=UserUpdate)
-@auth_guard("user")
-def modify_user(request: Request):
-    """
-    Update a user.
-    """
-    try:
-        # crud_users.CRUD_users.modify(id)
-        print(request.attach_user)
-    except HTTPException:
-        pass
-
-
-@router.delete("/{id}")
+@router.delete("/me")
 @auth_guard("user")
 def delete_user(request: Request):
-    """
-    Delete a user.
-    """
-    try:
-        # crud_users.CRUD_users.delete(id)
-        print(request.attach_user)
-        return {"Status": "OK"}
-    except HTTPException:
-        pass
+	"""_summary_
+
+	Args:
+		request (Request): _description_
+
+	Returns:
+		_type_: _description_
+	"""
+	user = User.assert_model_id(request.attach_user)
+	users.soft_delete(user)
+	return { "status": "OK" }