+++ backend/app/mongo/models/likes.py
@@ -1,12 +1,12 @@
 from pydantic import BaseModel, Field
+from typing_extensions import Self
+from mongo.schemas.likes import LikeUpdateComment, LikeUpdatePost
 
 
 class LikeBase(BaseModel):
     id: str = Field(alias="_id")
     id_user: str = Field(alias="userId")
     is_liked: bool
-    post_id: str = Field(alias="postId")
-    comment_id: str = Field(alias="commentId")
 
     class Config:
         schema_extra = {
@@ -18,3 +18,38 @@ class LikeBase(BaseModel):
                 "comment_id": 983,
             }
         }
+
+class LikeComment(LikeBase):
+    comment_id: str = Field(alias="commentId")
+
+
+class LikePost(LikeBase):
+    post_id: str = Field(alias="postId")
+
+
+class LikeToCreatePost(BaseModel):
+	userId: str
+	isLiked: bool
+	postId: str
+	
+	@classmethod
+	def assert_model(self, user_id: str, like_data: LikeUpdatePost) -> Self:
+		return self(
+			userId=user_id,
+			isLiked=like_data.is_liked,
+			postId=like_data.post_id,
+		)
+
+class LikeToCreateComment(BaseModel):
+	userId: str
+	isLiked: bool
+	commentId: str
+	
+	@classmethod
+	def assert_model(self, user_id, like_data: LikeUpdateComment) -> Self:
+		print(like_data)
+		return self(
+			userId=user_id,
+			isLiked=like_data.is_liked,
+			commentId=like_data.comment_id,
+		)