+++ backend/app/mongo/models/users.py
@@ -1,6 +1,12 @@
-from typing import Optional
+from typing import Dict, Optional
+from typing_extensions import Self
 from pydantic import BaseModel, EmailStr, Field as PydanticField
 from bson import ObjectId
+from mongo.schemas.users import UserCreate
+
+from core.security import (
+	get_password_hash
+)
 
 class PyObjectId(ObjectId):
     @classmethod
@@ -17,11 +23,28 @@ class PyObjectId(ObjectId):
 
 
 class User(BaseModel):
-    id: Optional[PyObjectId] = PydanticField(default_factory=PyObjectId, alias="_id")
-    username: str
+    id: Optional[PyObjectId]
+    username: Optional[str] = ...
+    profile_picture: Optional[str] = None
     email: EmailStr
     password: str
     isAdmin: bool = False
+    
+    @classmethod
+    def assert_model_id(self, user_data: Dict) -> Self:
+        if "_id" in user_data:
+            id = user_data["_id"]
+        else:
+            id = user_data["id"]
+
+        return self(
+            id=id,
+            username=user_data["username"],
+            email=user_data["email"],
+            password=user_data["password"],
+            isAdmin=user_data["isAdmin"],
+            profile_picture=user_data["profile_picture"]
+        )
 
     class Config:
         schema_extra = {
@@ -35,3 +58,19 @@ class User(BaseModel):
                 "refresh_token": "rMRCZVOfcK"
             }
         }
+
+
+class UserToCreate(BaseModel):
+    username: str
+    email: EmailStr
+    password: str
+    isAdmin: bool = False
+    profile_picture: str = None
+    
+    @classmethod
+    def assert_model(self, user_data: UserCreate) -> Self:
+        return self(
+            username=user_data.username,
+            email=user_data.email,
+            password=get_password_hash(user_data.password),
+        )